[Console] Extract `TextWriter` API
